+++ packages/lynx_view_android/android/src/main/kotlin/com/geektree0101/lynx_view_android/LynxPlatformView.kt
@@ -31,7 +31,9 @@ internal class LynxPlatformView(
     creationParams: Map<String?, Any?>?,
 ) : PlatformView, MethodChannel.MethodCallHandler {
 
-    private val lynxView: LynxView = LynxViewBuilder().build(context)
+    private val lynxView: LynxView = LynxViewBuilder()
+        .also { XElementSupport.addBehaviorsIfPresent(it) }
+        .build(context)
     private val channel = MethodChannel(binaryMessenger, "$INSTANCE_CHANNEL_PREFIX$viewId")
     private val mainHandler = Handler(Looper.getMainLooper())